+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2019-02-07
+### Added
+- Fréchet distance now supports any minkowski distance!
+### Changed
+- Fréchet distance now supports N-D data!
+- Add notes to Fréchet and DTW about support of N-D data in readme.
+### Removed
+- euc_dist function has been removed
+
 ## [0.2.3] - 2018-12-10
 ### Changed
 - fixed image link in README.rst for pypi.org